+++ test/dummy/app/javascript/application.js
@@ -1,10 +1,10 @@
 import "@hotwired/turbo-rails"
 import { Application } from "@hotwired/stimulus"
-import JanelaDashboardController from "janela/dashboard_controller"
+import JanelaFrameController from "janela/frame_controller"
 import JanelaChartController from "janela/chart_controller"
 
 const application = Application.start()
-application.register("janela--dashboard", JanelaDashboardController)
+application.register("janela--frame", JanelaFrameController)
 application.register("janela--chart", JanelaChartController)
 
 window.Stimulus = application